added a movementControl Call to main

+++ movement.c
@@ -4,7 +4,7 @@
 extern void setMotorSettings(uint8_t motorADirection, uint8_t motorAPWM,
                              uint8_t motorBDirection, uint8_t motorBPWM);
            
-void movementControl(int y_stick, int x_stick){
+void movementControl(uint16_t y_stick, uint16_t x_stick){
     uint8_t motorADirection = 0;
     uint8_t motorBDirection = 0;
     uint8_t motorAPWM = 0;
@@ -61,7 +61,7 @@ void movementControl(int y_stick, int x_stick){
         motorBPWM = 0;       // Default PWM
     }
 
-    
+
     if (motorAPWM < 30){
         motorAPWM = 30;
     } else if (motorAPWM > 100){
